The Department of Energy said Wednesday night it was canceling 321 awards worth $7.56 billion that were largely focused on clean energy.
The agency hasnt publicly released a list of the affected projects and, at the time of publication, it had not provided one to TechCrunch. According to E&E News and Heatmap, which have obtained the list, the majority of the cuts have hit states that went for Kamala Harris in the last presidential election, though some were in “red” states that voted for President Trump.