The U.S. government shutdown could stifle deal flow, freeze visa processing for workers, and cause other problems for startups and the broader tech sector, especially if it lasts longer than a week, according to experts who spoke to TechCrunch.
The U.S. government shutdown, which began Tuesday, is the first one in seven years. The unpredictability of the Trump administration coupled with a politically entrenched Congress makes it hard to predict when the shutdown will end. Out of eight shutdowns since 1990, four have happened during Trump administrations; the last one lasted 35 days, the longest in modern history.