Amazon will restart its drone delivery service  in Arizona beginning Friday as two federal agencies continue to investigate a crash that occurred earlier this week.
Amazon suspended operations in the West Valley of the Phoenix Metro area — currently its only commercial market — following a crash Wednesday. Two of the companys Prime Air delivery drones collided with the boom of a crane near its same-day site in Tolleson, Arizona, which sent them crashing to the ground.