Big tech
Dutch court orders Meta to change its Facebook and Instagram timelines

This is the latest action taken to reign in big tech under the EU's Digital Services Act.

Andre Revilla

Contributing Reporter

Andre RevillaContributing Reporter

Thu, October 2, 2025 at 6:51 PM UTC

2 min read

NurPhoto via Getty Images

A court in the Netherlands has ordered Meta to change Facebook and Instagram's timelines, after finding that the element ran afoul of the European Union's Digital Services Act (DSA). As reported by Reuters, the Dutch court said that the company needs to provide users with simpler options — namely ones that don't rely on an algorithm.

"People in the Netherlands are not sufficiently able to make free and autonomous choices about the use of profiled recommendation systems," the court said in its decision. It ruled that the timeline must honor a user's choice of chronological order or other non-profiled options, instead of reverting to the algorithm-driven version whenever a user closes and reopens either app.

The case was brought by Bits of Freedom, a Dutch digital rights group. "It is unacceptable that a few American tech billionaires can determine how we view the world," said the group's spokesperson, Maartje Knaap.

Advertisement

Advertisement

Advertisement

Meta said it will appeal the decision, and that these DSA issues should be handled by the European Commission and other EU regulators, not by the courts of individual nations. "Proceedings like this threaten the digital single market and the harmonized regulatory regime that should underpin it," a Meta spokesperson said. Meta faces a potential fine of $117,450 for every day it fails to comply with the court's order, up to a maximum of $5.8 million.

The DSA has been a common thorn in the side of big tech companies since its approval in 2022. The European Commission has levied hundreds of millions of dollars in fines against the likes of Apple, Meta and Alphabet for violations of the DSA. The regulations have also been used to effect certain changes on these platforms in the name of privacy, data security and the protection of minors.

About our ads