Techcrunch:
Welcome to YC Arena, which is not a top secret fight club for Y Combinator founders but rather a suite of games that give you a vague sense of what its like to be a partner at YC.

Created by a student in Berlin, YC Arenas YC Partner Simulator game shows you a publicly available pitch video from a company that applied to YC, along with the year of their application. You click &#8220;accept&#8221; or &#8220;reject,&#8221; and then find out if you made the same choice as YC.

Image Credits:YCArena

Its a lot harder than it looks. YC is estimated to accept around 1% of applicants, and at a certain point, some luck is required to really capture a partners eye &#8212; maybe your pitch is the first that a partner watched after a very rejuvenating coffee break, or maybe your company is the last video on the list and everyones tired.