Techcrunch:
OpenAIs Sora app for AI videos is a viral hit, despite being invite-only for now and limited to users in the U.S. and Canada at launch. On its first day, Sora saw 56,000 downloads, and is now ranked as the No. 3 Top Overall app on Apples U.S. App Store, according to new data from app intelligence provider Appfigures. On Friday, October 3, it became the No. 1 app on the U.S. App Store.

That puts it ahead of Google Gemini and OpenAIs ChatGPT.