Techcrunch:
This week, the Department of Energy canceled nearly $8 billion worth of awards — a move touted by the Trump administration as an effort to protect fossil fuels at the expense of renewables. But documents obtained by TechCrunch show that the reality is more complex than that simple message.

The agency has not released a list of the canceled awards, but TechCrunch has obtained a copy and has analyzed the 321 contracts that the DOE is seeking to undo.