Engadget: Deals
The Roku Streaming Stick Plus drops to only $29 in this Prime Day deal

This travel-size streaming device was our top pick for free and live content.

Andre Revilla

Contributing Reporter

Andre RevillaContributing Reporter

Updated Fri, October 3, 2025 at 7:01 PM UTC

Jeff Dunn for Engadget

The Roku Streaming Stick Plus is on sale for just $29. That's a discount of 27 percent and the lowest we've ever seen it.

Roku has held the top spot in the TV OS market for years thanks to its user-friendly interface, an affordable range of streaming devices and its own lineup of TVs. We picked the Streaming Stick Plus as the best streaming device for free and live content, thanks in large part to The Roku Channel app that accompanies it. The Roku Channel features over 500 free TV channels with live news, sports coverage and a rotating lineup of TV shows and movies.

Roku

Roku Streaming Stick Plus

$29$40Save $11

On sale for almost 30 percent off.

$29 at Amazon

In our hands-on review of the Roku Streaming Stick Plus, we thought it was perfect for travel thanks to its small size and the fact that it can be powered by your TV's USB port, nixing the need for a wall adapter. Menu navigation and opening or closing apps won't happen at quite the same speeds as more expensive streamers, but it's quick enough for what is ultimately a pretty low-cost option. The Wi-Fi range on this one is also weaker than Roku's pricier devices, but unless you are placing it exceedingly far from your router, it shouldn't be an issue.

Advertisement

Advertisement

The Roku Streaming Stick Plus supports both HD and 4K TVs, as well as HDR10+ content. It doesn't support Dolby Vision, however; for that you'll need to upgrade to Roku's Streaming Stick 4K or Roku Ultra. It comes with Roku's rechargeable voice remote with push-to-talk voice controls. Roku's remote can also turn on your TV and adjust the volume while you're watching.

If you've been thinking about getting a Roku device, or you already love the platform and want a compact and convenient way to take it with you when you travel, then this sale provides a great opportunity. If you'd prefer a more powerful and slightly more stationary device, the Roku Ultra is on sale for $79 right now, too.

Roku Ultra

$79$100Save $21

See at Amazon

Shop Prime Day deals: Early October Prime Day 2025 deals | Early Prime Day tech deals

About our ads