Techcrunch:
Drivers for ride-hailing apps like Uber and Lyft will soon have the right to unionize in California as independent contractors, thanks to a bill signed Friday by Governor Gavin Newsom.

This is part of a larger deal between lawmakers, unions, and ride-hailing companies, resulting in the passage of separate bills supporting lower insurance requirements for Uber and Lyft, along with union rights for their drivers. When the deal was first announced in August, Newsom described it as an “historic agreement between workers and business that only California could deliver.”