Techcrunch:
New PitchBook data illustrates how dramatically AI is dominating startup investment, with 2025 on-track to become the first year when AI accounts for more than half of all VC money invested.

PitchBook reports that VCs have poured $192.7 billion into the industry so far this year, out of a total $366.8 billion, according to Bloomberg. In the most recent quarter, AI accounted for 62.7% of the money invested by U.S. VCs, and for 53.2% of money invested by global firms.