Techcrunch:
OpenAI has acquired Roi, an AI-powered personal finance app. In keeping with a recent trend in the AI industry, only the CEO is making the jump.

Chief executive and co-founder Sujith Vishwajith announced the acquisition on Friday, and a source familiar with the matter told TechCrunch he is the only one of Rois four-person staff to join OpenAI. Terms of the deal were not disclosed. The company will wind down operations and end its service to customers on October 15.