Techcrunch:
Naveen Rao, who was formerly the head of artificial intelligence at Databricks, is in talks to raise $1 billion at a $5 billion valuation for a new startup called Unconventional, Inc., thats building a new type of computer, according to four people familiar with discussions.

Andreessen Horowitz has agreed to lead the investment, and Lightspeed and Lux Capital are also participating in the deal, two of the sources said, although none of those VC firms responded to our request for comment. Databricks is also said to be investing in Raos new company, Bloomberg reported last month.