Engadget: Gaming
Kingmakers, the medieval battle game with modern weapons, has been delayed

The game was supposed to launch in Early Access just days from now.

Cheyenne MacDonald

Weekend Editor

Cheyenne MacDonaldWeekend Editor

Sat, October 4, 2025 at 9:19 PM UTC

2 min read

Redemption Road Games

Redemption Road's absolutely bonkers-looking medieval shooter, Kingmakers, was slated to launch in Early Access on October 8, but now its release has been pushed back with no new date in sight. The developers posted an update on Steam to say that the scheduled launch, just days away, "will no longer be possible," going on to explain that they need "a bit more time on content polish before we feel good about charging money for it."

A statement posted by the developers of Kingmakers announcing that the game is delayed

(Redemption Road Games)

Kingmakers has generated a fair amount of hype since it was announced last year, and it sits among the top wishlisted titles on Steam. The game sends players "back in time to a war-torn medieval era with a vast arsenal of modern weapons," from guns to tanks. And if you've seen the trailers, you know it's not just the concept that's bananas — the gameplay we've been shown so far is completely over the top. In its statement about the delay, the team said that Kingmakers is "an incredibly ambitious, uncompromising game, and we don't want to cut any planned features, for the sake of getting it out the door earlier."

"With Kingmakers, we set out to push the Unreal Engine 4 codebase to its absolute limits, while still providing true 60fps to midrange PCs, without the need for fake frames," the team wrote. "We are an 80% engineering team, who got into this business to push technological barriers. We currently have tens of thousands of soldiers, each with AI and pathfinding that rivals what you'd expect from a AAA third person shooter. When you walk away from a battle, it continues to play out. Nothing is faked."

The developers haven't provided any update on what the timeline looks like now beyond the fact that the game won't be ready for October 8. But, they added, "We will be presenting a half hour long deepdive on Kingmakers gameplay very soon, with a comprehensive overview of everything we've been working on."

About our ads