Engadget: Entertainment
What to read this weekend: Ghost stories to kick off October

Check out Bora Chung's Midnight Timetable and the new comic series, It Killed Everyone But Me.

Cheyenne MacDonald

Weekend Editor

Cheyenne MacDonaldWeekend Editor

Updated Sat, October 4, 2025 at 6:29 PM UTC

Here are some recently released titles to add to your reading list. This week, we read Bora Chung's Midnight Timetable: A Novel in Ghost Stories, and started a new series from Mad Cave Studios, It Killed Everyone But Me.

Algonquin Books

Midnight Timetable

"Spooky season" is all year round in my world, but I still can't resist the urge to plunge even deeper into all things eerie when that autumn chill starts creeping in. And what better way to get into that vibe than with a collection of ghost stories? Midnight Timetable is a short book composed of seven interconnected stories as told to a new nighttime employee of a facility known as The Institute. The rooms of The Institute are home to a menagerie of haunted objects and beings, each with its own ghastly backstory, and strange things tend to occur on the premises.

This was my introduction to Bora Chung's work and it absolutely hooked me. It's a really quick read with just the right amount of creep factor.

$14 at Amazon

Mad Cave

It Killed Everyone But Me #1

This has been sitting on my To Be Read list for weeks, but October seemed like the perfect time to finally dive into a fresh slasher story. In It Killed Everyone But Me, we meet a Final Girl, Sutton, decades after the violent tragedy she survived. She's the mom of a teenager now, works in real estate and is going through all the motions of a "normal" life. But when someone approaches her about a new case that appears to have similarities to hers, Sutton is forced to revisit her past. “This story is about somebody literally burying their trauma away and ignoring it and then having to dig it up 30 years later and deal with it,” writer Ryan Parrott said in a recent interview with AIPT. He added that it is, "in some ways, a cautionary tale about how you cant just pretend something didnt happen and be OK."

This first issue is a strong start to the series, and leaves you hungry for more. There's a lot we still don't know — like what really happened back then, when all of Sutton's friends were killed during a weekend getaway.

$5 at Amazon

About our ads