Techcrunch:
AI coding tools are getting better fast. If you dont work in code, it can be hard to notice how much things are changing, but GPT-5 and Gemini 2.5 have made a whole new set of developer tricks possible to automate, and last week Sonnet 4.5 did it again.

At the same time, other skills are progressing more slowly. If you are using AI to write emails, youre probably getting the same value out of it you did a year ago. Even when the model gets better, the product doesnt always benefit — particularly when the product is a chatbot thats doing a dozen different jobs at the same time. AI is still making progress, but its not as evenly distributed as it used to be.