Techcrunch:
The tie between startups and the U.S. government have strengthened in recent years, a shift buoyed by an interest in using AI, automation, space, robotics, and climate tech for defense. And while that has provided another welcome path to capital, the relationship is getting complicated.

A growing share of startups have the U.S. government as customers, or are aiming for permits and defense-related contracts. When the government is operational, that connection can provide a needed boost and revenue to startups. But when the government ceases to function, as it did starting October 1, those close ties can stifle or even halt progress for startups.