Engadget: Transportation
Uber and Lyft drivers in California are able to unionize under new law

The 800,000-plus rideshare drivers in the state will be able to collectively bargain for better conditions.

Cheyenne MacDonald

Weekend Editor

Cheyenne MacDonaldWeekend Editor

Sun, October 5, 2025 at 5:13 PM UTC

1 min read

REUTERS / Reuters

California governor Gavin Newsom signed legislation on Friday that grants rideshare drivers in the state the right to unionize. It's the second state to grant organizing rights to rideshare drivers, who are independent contractors, following the passage of a similar law in Massachusetts in 2024. There are over 800,000 rideshare drivers in California, and the bill that was just signed into law "establishes a clear legal framework for union certification, bargaining processes and enforcement," according to a press release from the office of Assemblymember Buffy Wicks.

This means drivers working for companies like Uber and Lyft will be able to collectively bargain for better pay, benefits and working conditions. Under the terms of the law, driver organizations will be able to apply for union recognition starting in May 2026 as long as they have support from at least 10 percent of active rideshare drivers in the state. The organization would then need support from at least 30 percent of active drivers to begin bargaining on their behalf.

As part of a deal made in September, Newsom also signed a measure that reduces the insurance coverage requirements for Uber and Lyft in the case of accidents caused by uninsured drivers, Associated Press reports.

About our ads