Engadget: Deals
This Roomba robot vacuum is on sale for $150 for Prime Day

Pick up the entry-level model for 40 percent off.

Lawrence Bonk

Contributing Reporter

Lawrence BonkContributing Reporter

Updated Sun, October 5, 2025 at 5:01 PM UTC

iRobot / Engadget

Robot vacuums are great items to look for during events like October Prime Day. They're usually hundreds of dollars off, so you can save a ton if you're buying one as a gift or you want to upgrade an aging robovac you have at home already. One of the best deals this time around is on the iRobot Roomba 104, which is 40 percent off and on sale for only $150.

This is a newer version of the unit that topped our list of the best budget robot vacuums. It's an entry-level robovac that gets the job done. The cleaning motor is fairly powerful and it ships with a multi-surface brush and an edge-sweeping brush. The vacuum uses LiDAR to map a home and to help it avoid obstacles when cleaning.

iRobot

iRobot Roomba 104 Robot Vacuum

$150$250Save $100

This is 40 percent off.

$150 at Amazon

It's also been equipped with specialized sensors to prevent falling down stairs. Steps are the natural enemy of all robot vacuums, except maybe this one. The Roomba 104 integrates with the company's proprietary app, which allows for custom cleaning schedules and the like. The robot can also be controlled via voice assistant and boasts compatibility with Siri, Alexa and Google Assistant.

Advertisement

Advertisement

The vacuum will automatically head to the charger for some juice when running low, which is nice. The battery lasts around 200 minutes per charge, which is a decent enough metric for a budget-friendly robovac. The only downside here? This is just a vacuum. It doesn't mop and it doesn't come with a dedicated debris canister.

Shop Prime Day deals: Early October Prime Day 2025 deals | Early Prime Day tech deals | Best credit card for Amazon

About our ads