Engadget: Deals
Prime Day Apple deals include the iPad A16 for $280

That's a record-low price.

Lawrence Bonk

Contributing Reporter

Lawrence BonkContributing Reporter

Tue, October 7, 2025 at 8:27 AM UTC

Engadget

The entry-level Apple iPad with the A16 chip is on sale for $280 as a Prime Day promotion. This is a discount of around $70 and represents a record-low price for the 11-inch tablet. The deal is available in multiple colorways.

This tablet made our list of the best iPads and we particularly recommend it for the budget-conscious. It may lack some of the bells and whistles of its more expensive cousins, but it still gets the job done. It's fast enough for most common tasks and the battery life is solid, at 10 hours per charge. It also ships with 6GB of RAM, which is a decent enough metric.

Apple

Apple iPad with an A16 chip

$280$349Save $69

This is a record-low price.

$280 at Amazon

The tablet integrates with Wi-Fi 6 and features a USB-C connector, along with a pair of 12MP cameras. There's also a Touch ID sensor, which can be used to unlock the device and to make payments.

Advertisement

Advertisement

This is the most basic tablet in Apple's lineup, so it lacks some features. It doesn't integrate with the Apple Pencil Pro or the Magic Keyboard. The display is also not quite as gorgeous as the iPad Air or Pro, but it's still a Liquid Retina panel.

Shop Prime Day deals: Early October Prime Day 2025 deals | Early Prime Day tech deals | Best credit card for Amazon

Advertisement

About our ads