Techcrunch:
Zendesk announced Wednesday at its AI summit a string of LLM-driven products meant to reshape the companys reliance on human technicians.

The center of the new features is an autonomous support agent that Zendesk believes will solve 80% of support issues without human intervention. That system will be supplemented by a co-pilot agent, which will assist human technicians with the remaining 20% of issues, as well as an admin-layer agent, a voice-based agent, and an analytics agent.