Engadget: Transportation
EVs

Ferrari's first EV is coming next year with big speed, big sound and a Jony Ive design

All the details we know about Ferrari's next new model.

Tim Stevens

Contributor

Tim StevensContributor

Thu, October 9, 2025 at 8:30 AM UTC

6 min read

The chassis, battery pack and modules for the Ferrari Elettrica.

(Ferrari)

Sound is a complicated topic in cars. Some cars are good because they're very loud. Some cars are good because they're very quiet. A select few cars can be counted in both categories. Really, only Hyundai's Ioniq 5 N has managed to check both boxes, but if Ferrari's promises about its first EV are to be believed, we may soon have another entrant.

At an event at the company's headquarters in Maranello, Italy, Ferrari unveiled the first proper details about its first EV, the Elettrica, which is set to fully debut next year. We're missing a few key details (such as what it looks like), but after spending a day talking to the various engineers behind this machine, I can finally reveal what makes this car move and how it might sound, too.

Power is a key metric for any Ferrari, and the Elettrica has plenty of that. Four electric motors, one per wheel, combine to generate a figure somewhere over 1,000 horsepower. That would put it close to the power of the company's new F80 supercar, but in an era of 2,000-hp EVs, the Elettrica's output isn't exactly stunning.

Advertisement

Advertisement

Advertisement

This is something even Ferrari's chief product development officer, Gianmaria Fulgenzi, admitted: "It's very easy and simple to create power in an electric engine. It's not difficult." Cornering, he said, is what it's really all about. He called current performance EVs "elephants," capable of going quickly in a straight line but failing miserably at being compelling machines to drive.

That's partly why the Elettrica has four motors. That way, the car's traction and stability management systems can modulate power to maximize grip at each tire individually, rather than relying on one or two motors with differentials to try to manage wheelspin.But it goes deeper than that. The Elettrica can also steer each rear wheel independently, more or less on the inside or outside of the turn to help manage (or perhaps even induce) oversteer.

The front motors and inverter on the Ferrari Elettrica.

(Ferrari)

Finally, the Elettrica will have active suspension, a novel type that debuted on the Purosangue SUV and is also used on the F80 supercar. Ferrari's dampers replace traditional valves and oil with an electric motor that provides the resistance. This means, again, the damping and ride height of the car can be adjusted independently at each corner, nearly instantly.

All that, Ferrari CEO Benedetto Vigna said, will result in an EV that drives like none other on the road: "We want to show that we are able to harness any technology in a unique way." Making a compelling EV, a type of car that many decry as too samey, would be a bold way to do it.

Advertisement

Advertisement

Advertisement

One key factor to this is a unique sound. Ferrari engineers knew that they needed to come up with some sort of acoustic accompaniment for the Elettrica, but they also knew it had to be authentic. "We didn't create a fake sound, like a starship or something like that," Fulgenzi said. "We wanted to have exactly the sound of the electric engine."

Ferrari came up with a solution similar to the one Porsche used for the Taycan, but it executed it in a novel way. In the Taycan, Porsche sampled and remixed the sounds of the car's electrical components. For the Elettrica, Ferrari actually installs an accelerometer inside the casing of the rear electric motors. That sensor picks up the harmonic resonance and runs it through a digital amplifier to create a sound that isn't a sample — its actually coming directly from those motors.

Fulgenzi likens it to an amp on an electric guitar, picking up the near-inaudible vibration of the strings and turning it into (potentially ear-splitting) sound. That's perhaps a bit aspirational, but the proof of course will be in the listening, something that Ferrari sadly hasn't let me (or anyone else) do just yet.

The rear motor and inverter for the Ferrari Elettrica.

(Ferrari)

They're also not showing anyone what the thing looks like, but it seems safe to expect that this machine will bear a shape and design unlike any Ferrari that's come before. Thats because Ferrari is, for the first time, working with design house LoveFrom, better known as the house that Jony Ive built after leaving Apple in 2019 and then sold to OpenAI for $6.5 billion.

Advertisement

Advertisement

Advertisement

LoveFrom has worked on a fair few projects since its inception, but this will be the company's first car. Given that, we have absolutely nothing to go on when it comes to visual style. But if I had to guess, I'd say something with a minimalist vibe is on the menu based on Ive's previous work. (And also his own car collection, which features a lot of classic bits of iconic British motoring.)

What we do know is that the Elettrica will have four doors and four seats. No, then, this won't be a proper sports car, instead sitting somewhere between grand tourer and SUV. It'll be something designed with a bit of practicality and comfort, able to waft calmly and quietly if you like. Or, with the twist of a few knobs on the steering wheel, it's able to tighten up and get loud when you're feeling more rowdy.

With somewhere north of 330 miles of range, it should make for a decent tourer, too. That comes from a 122 kWh (gross) battery pack that's split up into 15 modules, each containing 14 NMC pouch cells. The cells are initially sourced from SK On, but Ferrari made it clear they're not tied to that supplier.

Indeed, the company hopes to evolve the battery pack design and chemistry over the years, keeping some version of it available for decades to come. This is, after all, a Ferrari, a machine that ostensibly will just keep getting more valuable as it ages. That, though, is a challenge currently stymying other makers of premium EVs.

Can Ferrari make a truly desirable, truly premium EV? Having Jony Ive's name attached certainly doesn't hurt. We'll all have to wait until early next year to see how it looks and how it sounds. I personally won't get too excited about this car until then, but I am very glad that it even exists. Back in 2011, Luca Di Montezemelo, then-CEO of Ferrari, told me that Ferrari would never make an EV. I was skeptical of his statement back then, and 14 years later, I'm glad to see the company finally coming around.

Advertisement

About our ads