Techcrunch:
Venture capital firm Andreessen Horowitz (a16z) has pushed back on local media reports suggesting it plans to open an office in India, with a general partner calling the coverage &#8220;entirely fake news.&#8221;

Earlier on Thursday, Indian media outlets reported that a16z was preparing to establish a physical presence in the country by setting up an office in Bengaluru. The reports, which cited unnamed sources, also said the firm was in the process of hiring a local partner.