Engadget: Gaming
Nintendo

Nintendo just surprise-dropped a Japan-only SNES Mario game onto the Virtual Console

It actually uses the Joy-Con mouse functionality on Switch 2.

Lawrence Bonk

Contributing Reporter

Lawrence BonkContributing Reporter

Thu, October 9, 2025 at 6:35 PM UTC

1 min read

Nintendo

Nintendo just released an odd little game onto the SNES Virtual Console that was previously exclusive to Japan. The simply-titled Mario & Wario is a point-and-click puzzle game that finds the plumber on the search for his brother after his money-obsessed doppelganger drops a bucket on his head. That's the actual plot.

Players lead the bucket-headed Mario through 100 levels, each of which are filled with point-and-click puzzles set in a sidescrolling world. The original game was controlled via mouse and this Virtual Console release is no different. Switch owners can simply plug in a USB mouse to get going but Switch 2 owners have a niftier option.

You may remember that the new console's Joy-Con controllers can flip over and be used as a mouse. Mario & Wario is another use case for this concept, after Drag x Drive and a minigame in Nintendo Switch 2 Welcome Tour. The game is available right now for Nintendo Switch Online members.

That's not the only title that hit Virtual Console today. The SNES platformer Bubsy in Claws Encounters of the Furred Kind is available to play, as is the SNES fighter Fatal Fury Special. That one includes characters from both the original game and Fatal Fury 2.

Advertisement

About our ads