Techcrunch:
It takes a lot of computing power to run an AI product — and as the tech industry races to tap the power of AI models, theres a parallel race underway to build the infrastructure that will power them. On a recent earnings call, Nvidia CEO Jensen Huang estimated that between $3 trillion and $4 trillion will be spent on AI infrastructure by the end of the decade — with much of that money coming from AI companies. Along the way, theyre placing immense strain on power grids and pushing the industrys building capacity to its limit.

Below, weve laid out everything we know about the biggest AI infrastructure projects, including major spending from Meta, Oracle, Microsoft, Google, and OpenAI. Well keep it updated as the boom continues and the numbers climb even higher.

Microsofts $1 billion investment in OpenAI

This is arguably the deal that kicked off the whole contemporary AI boom: In 2019, Microsoft made a $1 billion investment in a buzzy non-profit called OpenAI, known mostly for its association with Elon Musk. Crucially, the deal made Microsoft the exclusive cloud provider for OpenAI — and as the demands of model training became more intense, more of Microsofts investment started to come in the form of Azure cloud credit rather than cash.