Techcrunch:
There are a number of iPad apps that can help you explore and express your creativity. Although the iPad started off as a simple device that could be used to stream content or browse the web on the go, Apple has essentially turned its iPads into powerful machines that can be used to do things like create digital art and edit videos.

Weve compiled a list of some of the best iPad apps for creativity that are available on the App Store.