Engadget: Big tech
The FCC is trying to make it easier for internet providers to charge hidden fees

A new proposal would weaken Biden-era transparency rules.

Lawrence Bonk

Contributing Reporter

Lawrence BonkContributing Reporter

Fri, October 10, 2025 at 3:53 PM UTC

2 min read

Brandon Bell via Getty Images

The FCC has submitted a proposal that would make it harder for consumers to receive itemized bills with accurate information from their ISPs, as spotted by The Verge. The draft would revise previous "unnecessary" requirements on the grounds that a fact-based list of charges "may confuse customers."

This comes as a response to complaints by ISPs over a Biden-era transparency rule that went into effect back in 2024. Compliance with this rule mandates that ISPs make it easier for consumers to compare prices and avoid hidden fees. This typically comes in the form or something called a "nutrition label", which shows speeds, fees and data allowances for internet plans.

The FCC is planning on examining ways to eliminate "any other label requirements that are unduly burdensome and provide minimal benefit to consumers." Getting an actual list of what we are paying $100 for each month doesn't seem like a minimal benefit to me.

Advertisement

Advertisement

Advertisement

The regulatory agency isn't looking to ditch these nutrition labels entirely, at least not yet, but is seeking to eliminate the requirements to itemize location-based fees. "We propose to eliminate the requirement that providers itemize discretionary, recurring monthly fees that represent costs they choose to pass through to consumers and which vary by consumer location," the draft proposal said.

It also targets requirements to make this data readily available to all consumers. The proposal has language that would stop forcing ISPs to make this information multi-lingual and to make it available both online and to customers on the phone. It also seeks to decouple these labels from customer account portals. We aren't sure where these nutrition labels would continue to live.

FCC head Brendan Carr wrote a cheeky blog about the proposal that seemed more interested in discussing pumpkin spice lattes over ISP regulation. He said that the proposal seeks to "separate the wheat from the chaff" so that consumers can "get quick and easy access to the information they want and need to compare broadband plans." The language of the proposal seems to indicate the exact opposite of that sentiment, but that's politics for ya.

A vote on the proposal is scheduled for October 28. This is all part of Carr's "Delete, Delete, Delete" initiative, which seeks to appease President Trump by removing as many government regulations as possible.

Advertisement

About our ads