Techcrunch:
Apple appears to be winding down support of Clips, with the company removing Clips from the App Store and saying it will no longer be making any updates.

In a support page on the Apple website, the company says that as of October 10, Clips is no longer available for new users to download, but existing users can continue to use the app on current or earlier versions of iOS and iPadOS. Existing users can also re-download the app from their Apple account if needed.