Engadget: Apple is winding down Clips, its forgotten video-editing app
Existing users will still have access for now, but it's no longer available for new downloads.

Cheyenne MacDonald

Weekend Editor

Cheyenne MacDonaldWeekend Editor

Sat, October 11, 2025 at 9:09 PM UTC

1 min read

Apple

It seems to be the end of the road for Apple's Clips app. The company said on a support page that it has stopped updating the video-editing app, and new users can no longer download it from the App Store. Existing users on iOS and iPadOS will continue to have access for the time being, but the company hasn't said how long that will last. "If you previously downloaded the Clips app, you can still redownload it from your Apple account in the App Store," the support page notes.

Clips was introduced in 2017 and offers a host of editing tools geared toward creating videos for social media. But Clips never really caught on, and many of the features it boasts are pretty much standard fare on the social apps now. Apple has recommended that Clips users save their  videos to their photo libraries or elsewhere, so if you are one of the few users hanging on, you should probably go ahead and do that before it's too late.

Advertisement

About our ads