Techcrunch:
Goldman Sachs has agreed to acquire Industry Ventures, a 25-year-old, San Francisco-based investment firm with $7 billion in assets under management, CNBC was first to report on Monday. The deal underscores the growing importance of secondary markets and buyouts as traditional venture exits remain sluggish.

The investment bank is paying $665 million in cash and equity, with up to $300 million more tied to the firms performance through 2030, according to a release from Goldman. The deal is expected to close in the first quarter of next year, and all 45 Industry Ventures employees are expected to join Goldman.